+++ CHANGELOG.md
@@ -81,6 +81,8 @@
 - Added `title` and `description` as fields in the response of `Get Channel Chat Badges` and `Get Global Chat Badges`
 - Added `Get AutoMod Settings` and `Update AutoMod Settings` endpoints
 - Added `Get Unban Requests` and `Resolve Unban Requests` endpoints
+- Added `Get Moderated Channels` endpoint
+- Added `Get User Emotes` endpoint
 
 ### Fixed